Adding the basic function
When the user enters a string message, it will repate the input and then all the postive words corrsponding with the lietter of the word. the posstive words will be placed under the postive message.

I just finished the actual code in such short time. so im actually proud. the only thing i i used ai for was the looping sections, so not too bad